# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.SubmitTier1Request:
    properties:
      bvn:
        type: string
      date_of_birth:
        type: string
      display_username_on_p2p:
        type: boolean
      first_name:
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        maxLength: 100
        minLength: 2
        type: string
      phone_number:
        maxLength: 20
        minLength: 10
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - bvn
    - date_of_birth
    - first_name
    - last_name
    - phone_number
    - username
    type: object
  handler.SubmitTier2Request:
    properties:
      nin:
        type: string
    required:
    - nin
    type: object
...
  models.KYCStatus:
    enum:
    - pending
    - verified
    - rejected
    type: string
    x-enum-varnames:
    - KYCStatusPending
    - KYCStatusVerified
    - KYCStatusRejected
  models.KYCStatusResponse:
    properties:
      address_verified:
        type: boolean
      bvn_verified:
        type: boolean
      next_tier:
        $ref: '#/definitions/models.KYCTier'
      nin_verified:
        type: boolean
      rejection_reason:
        type: string
      selfie_verified:
        type: boolean
      status:
        $ref: '#/definitions/models.KYCStatus'
      tier:
        $ref: '#/definitions/models.KYCTier'
    type: object
  models.KYCTier:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - KYCTierNone
    - KYCTier1
    - KYCTier2
    - KYCTier3
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready backend API for Kynettic platform with authentication,
    user management, and health monitoring
  title: Kynettic API
  version: "1.0"
paths:
...
  /users/me/kyc/status:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's KYC verification status and tier
      produces:
      - application/json
      responses:
        "200":
          description: KYC status retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.KYCStatusResponse'
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get KYC status
      tags:
      - kyc
  /users/me/kyc/tier1:
    post:
      consumes:
      - application/json
      description: Submit user details and BVN for Tier 1 verification
      parameters:
      - description: Tier 1 verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitTier1Request'
      produces:
      - application/json
      responses:
        "200":
          description: Tier 1 verification completed successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request or verification failed
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Submit Tier 1 KYC
      tags:
      - kyc
  /users/me/kyc/tier2:
    post:
      consumes:
      - application/json
      description: Submit NIN for Tier 2 verification
      parameters:
      - description: Tier 2 verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SubmitTier2Request'
      produces:
      - application/json
      responses:
        "200":
          description: Tier 2 verification completed successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request or verification failed
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Submit Tier 2 KYC
      tags:
      - kyc
  /users/me/kyc/tier3:
    post:
      consumes:
      - multipart/form-data
      description: Submit NIN, selfie image file, utility bill file, and address for
        Tier 3 verification. Files are uploaded to Cloudinary.
      parameters:
      - description: National Identification Number (11 digits)
        in: formData
        name: nin
        required: true
        type: string
      - description: Selfie image file (jpg, png)
        in: formData
        name: selfie_image
        required: true
        type: file
      - description: Utility bill document (jpg, png, pdf)
        in: formData
        name: utility_bill
        required: true
        type: file
      - description: Street address
        in: formData
        name: address_street
        required: true
        type: string
      - description: City
        in: formData
        name: address_city
        required: true
        type: string
      - description: State
        in: formData
        name: address_state
        required: true
        type: string
      - description: Country
        in: formData
        name: address_country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tier 3 verification completed successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid request or verification failed
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Submit Tier 3 KYC
      tags:
      - kyc
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"